refactor: :recycle: rename `RawTransaction` -> `Transaction`
BREAKING CHANGE: `RawTransaction` renamed to `Transaction`
