+++ Packages/com.careboo.unity-algorand-sdk/CareBoo.AlgoSdk/AlgoApi/Models/RawSignedTransaction.cs
@@ -9,7 +9,7 @@ namespace AlgoSdk
         : IEquatable<RawSignedTransaction>
     {
         [AlgoApiField("txn", "txn")]
-        public RawTransaction Transaction;
+        public Transaction Transaction;
 
         [AlgoApiField("sig", "sig")]
         public Signature Sig;
