--- Packages/com.careboo.unity-algorand-sdk/CareBoo.AlgoSdk/AlgoApi/Models/RawTransaction/RawTransaction.AssetFreeze.cs
@@ -1,26 +1,0 @@
-namespace AlgoSdk
-{
-    public partial struct RawTransaction
-    {
-        [AlgoApiField(null, "fadd")]
-        public Address FreezeAccount
-        {
-            get => AssetFreezeParams.FreezeAccount;
-            set => AssetFreezeParams.FreezeAccount = value;
-        }
-
-        [AlgoApiField(null, "faid")]
-        public ulong FreezeAsset
-        {
-            get => AssetFreezeParams.FreezeAsset;
-            set => AssetFreezeParams.FreezeAsset = value;
-        }
-
-        [AlgoApiField(null, "afrz")]
-        public Optional<bool> AssetFrozen
-        {
-            get => AssetFreezeParams.AssetFrozen;
-            set => AssetFreezeParams.AssetFrozen = value;
-        }
-    }
-}
