--- Packages/com.careboo.unity-algorand-sdk/CareBoo.AlgoSdk/AlgoApi/Models/RawTransaction/RawTransaction.AssetFreeze.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 2e680be51b9a1e449a98e4d2ec68703f
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
