--- Packages/com.careboo.unity-algorand-sdk/CareBoo.AlgoSdk/AlgoApi/Models/RawTransaction/RawTransaction.AssetTransfer.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 2d2dad6e68bf3b940a49b5a5f942f1f3
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
