--- Packages/com.careboo.unity-algorand-sdk/CareBoo.AlgoSdk/AlgoApi/Models/RawTransaction/RawTransaction.KeyRegistration.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: e875fda2f16ae8e48a9bdbce01c3f860
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
