+++ Packages/com.careboo.unity-algorand-sdk/CareBoo.AlgoSdk/AlgoApi/Models/Transaction/ITransaction.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 0deec979b6c286446a9c7e9ac6e20cbe
+guid: 2db637f313a91604a9cdb3d9ce0f7714
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
