+++ Packages/com.careboo.unity-algorand-sdk/CareBoo.AlgoSdk/AlgoApi/Models/Transaction/Signature/ISignature.cs
@@ -1,0 +1,6 @@
+namespace AlgoSdk
+{
+    public interface ISignature
+    {
+    }
+}
