+++ Packages/com.careboo.unity-algorand-sdk/CareBoo.AlgoSdk/AlgoApi/Models/Transaction/Signature/SignatureType.cs
@@ -1,0 +1,9 @@
+namespace AlgoSdk
+{
+    public enum SignatureType : byte
+    {
+        Sig,
+        MultiSig,
+        LogicSig
+    }
+}
