+++ Packages/com.careboo.unity-algorand-sdk/CareBoo.AlgoSdk/AlgoApi/Models/Transaction/SignedTransaction.cs
