+++ Packages/com.careboo.unity-algorand-sdk/CareBoo.AlgoSdk/AlgoApi/Models/Transaction/Transaction.KeyRegistration.cs
@@ -4,8 +4,50 @@ using Unity.Collections;
 
 namespace AlgoSdk
 {
-    public static partial class Transaction
+    public partial struct Transaction
     {
+        [AlgoApiField(null, "votekey")]
+        public Address VotePk
+        {
+            get => KeyRegistrationParams.VotePk;
+            set => KeyRegistrationParams.VotePk = value;
+        }
+
+        [AlgoApiField(null, "selkey")]
+        public VrfPubkey SelectionPk
+        {
+            get => KeyRegistrationParams.SelectionPk;
+            set => KeyRegistrationParams.SelectionPk = value;
+        }
+
+        [AlgoApiField(null, "votefst")]
+        public ulong VoteFirst
+        {
+            get => KeyRegistrationParams.VoteFirst;
+            set => KeyRegistrationParams.VoteFirst = value;
+        }
+
+        [AlgoApiField(null, "votelst")]
+        public ulong VoteLast
+        {
+            get => KeyRegistrationParams.VoteLast;
+            set => KeyRegistrationParams.VoteLast = value;
+        }
+
+        [AlgoApiField(null, "votekd")]
+        public ulong VoteKeyDilution
+        {
+            get => KeyRegistrationParams.VoteKeyDilution;
+            set => KeyRegistrationParams.VoteKeyDilution = value;
+        }
+
+        [AlgoApiField(null, "nonpart")]
+        public Optional<bool> NonParticipation
+        {
+            get => KeyRegistrationParams.NonParticipation;
+            set => KeyRegistrationParams.NonParticipation = value;
+        }
+
         public struct KeyRegistration
             : ITransaction
             , IEquatable<KeyRegistration>
@@ -134,16 +176,16 @@ namespace AlgoSdk
                 );
             }
 
-            public void CopyTo(ref RawTransaction rawTransaction)
+            public void CopyTo(ref Transaction transaction)
             {
-                rawTransaction.Header = header;
-                rawTransaction.KeyRegistrationParams = @params;
+                transaction.HeaderParams = header;
+                transaction.KeyRegistrationParams = @params;
             }
 
-            public void CopyFrom(RawTransaction rawTransaction)
+            public void CopyFrom(Transaction transaction)
             {
-                header = rawTransaction.Header;
-                @params = rawTransaction.KeyRegistrationParams;
+                header = transaction.HeaderParams;
+                @params = transaction.KeyRegistrationParams;
             }
 
             public bool Equals(KeyRegistration other)
