+++ Packages/com.careboo.unity-algorand-sdk/CareBoo.AlgoSdk/AlgoApi/Models/Transaction/TransactionType.cs
@@ -1,0 +1,17 @@
+using AlgoSdk.Formatters;
+
+namespace AlgoSdk
+{
+    [AlgoApiFormatter(typeof(TransactionTypeFormatter))]
+    public enum TransactionType : byte
+    {
+        None,
+        Payment,
+        KeyRegistration,
+        AssetTransfer,
+        AssetFreeze,
+        AssetConfiguration,
+        ApplicationCall,
+        Count,
+    }
+}
