+++ Packages/com.careboo.unity-algorand-sdk/CareBoo.AlgoSdk/AlgoApi/Models/Transaction/TransactionType.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 6ec819418eee66f47ac5f502c6a0c62b
+guid: 063ed16ac74ae0c4bb8224c0f120777f
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
