--- Packages/com.careboo.unity-algorand-sdk/CareBoo.AlgoSdk/Signature/ISignature.cs
@@ -1,6 +1,0 @@
-namespace AlgoSdk
-{
-    public interface ISignature
-    {
-    }
-}
