--- Packages/com.careboo.unity-algorand-sdk/CareBoo.AlgoSdk/Signature/SignatureType.cs
@@ -1,9 +1,0 @@
-namespace AlgoSdk
-{
-    public enum SignatureType : byte
-    {
-        Sig,
-        MultiSig,
-        LogicSig
-    }
-}
